requirements.txt
easydict==1.9
numpy==1.21.2
opencv_python==4.5.5.64
pillow==10.2.0
pyyaml==6.0
scipy==1.8.0
scipy==1.7.3
torch==1.8.2
torchvision==0.9.2
tqdm==4.64.0

